AC: separate SSD models adapters
